Move oj_clone
