+++ src/competitive_verifier/oj/verify/config.py
