+++ src/competitive_verifier/oj/verify/languages/java.py
@@ -2,10 +2,10 @@ import pathlib
 from logging import getLogger
 from typing import Any, Sequence
 
-import competitive_verifier_oj_clone.shlex2 as shlex
-from competitive_verifier_oj_clone.config import OjVerifyConfig
-from competitive_verifier_oj_clone.languages.models import LanguageEnvironment
-from competitive_verifier_oj_clone.languages.user_defined import UserDefinedLanguage
+import competitive_verifier.oj.verify.shlex2 as shlex
+from competitive_verifier.oj.verify.config import OjVerifyConfig
+from competitive_verifier.oj.verify.languages.models import LanguageEnvironment
+from competitive_verifier.oj.verify.languages.user_defined import UserDefinedLanguage
 
 logger = getLogger(__name__)
