+++ src/competitive_verifier/oj/verify/languages/nim.py
@@ -4,10 +4,13 @@ import pathlib
 from logging import getLogger
 from typing import Any
 
-import competitive_verifier_oj_clone.shlex2 as shlex
-from competitive_verifier_oj_clone.config import OjVerifyConfig
-from competitive_verifier_oj_clone.languages.models import Language, LanguageEnvironment
-from competitive_verifier_oj_clone.utils import read_text_normalized
+import competitive_verifier.oj.verify.shlex2 as shlex
+from competitive_verifier.oj.verify.config import OjVerifyConfig
+from competitive_verifier.oj.verify.languages.models import (
+    Language,
+    LanguageEnvironment,
+)
+from competitive_verifier.oj.verify.utils import read_text_normalized
 
 logger = getLogger(__name__)
