+++ src/competitive_verifier/oj/verify/languages/python.py
@@ -13,7 +13,10 @@ import importlab.environment
 import importlab.fs
 import importlab.graph
 
-from competitive_verifier_oj_clone.languages.models import Language, LanguageEnvironment
+from competitive_verifier.oj.verify.languages.models import (
+    Language,
+    LanguageEnvironment,
+)
 
 logger = getLogger(__name__)
