+++ src/competitive_verifier/oj/verify/list.py
@@ -1,17 +1,19 @@
-from competitive_verifier_oj_clone.config import OjVerifyConfig
-from competitive_verifier_oj_clone.languages.models import Language
+from competitive_verifier.oj.verify.config import OjVerifyConfig
+from competitive_verifier.oj.verify.languages.models import Language
 
 
 def get_dict(config: OjVerifyConfig) -> dict[str, Language]:
-    from competitive_verifier_oj_clone.languages.cplusplus import CPlusPlusLanguage
-    from competitive_verifier_oj_clone.languages.go import GoLanguage
-    from competitive_verifier_oj_clone.languages.haskell import HaskellLanguage
-    from competitive_verifier_oj_clone.languages.java import JavaLanguage
-    from competitive_verifier_oj_clone.languages.nim import NimLanguage
-    from competitive_verifier_oj_clone.languages.python import PythonLanguage
-    from competitive_verifier_oj_clone.languages.ruby import RubyLanguage
-    from competitive_verifier_oj_clone.languages.rust import RustLanguage
-    from competitive_verifier_oj_clone.languages.user_defined import UserDefinedLanguage
+    from competitive_verifier.oj.verify.languages.cplusplus import CPlusPlusLanguage
+    from competitive_verifier.oj.verify.languages.go import GoLanguage
+    from competitive_verifier.oj.verify.languages.haskell import HaskellLanguage
+    from competitive_verifier.oj.verify.languages.java import JavaLanguage
+    from competitive_verifier.oj.verify.languages.nim import NimLanguage
+    from competitive_verifier.oj.verify.languages.python import PythonLanguage
+    from competitive_verifier.oj.verify.languages.ruby import RubyLanguage
+    from competitive_verifier.oj.verify.languages.rust import RustLanguage
+    from competitive_verifier.oj.verify.languages.user_defined import (
+        UserDefinedLanguage,
+    )
 
     d: dict[str, Language] = {}
     d[".cpp"] = CPlusPlusLanguage(config=config)
