+++ tests/test_summary.py
@@ -5,10 +5,9 @@ from io import StringIO
 import pytest
 
 import competitive_verifier.summary as summary
-from competitive_verifier.models.result import FileResult, JudgeStatus
-from competitive_verifier.models.result import TestcaseResult as CaseResult
-from competitive_verifier.models.result import VerificationResult, VerifyCommandResult
-from competitive_verifier.models.result_status import ResultStatus
+from competitive_verifier.models import FileResult, JudgeStatus, ResultStatus
+from competitive_verifier.models import TestcaseResult as CaseResult
+from competitive_verifier.models import VerificationResult, VerifyCommandResult
 
 test_to_human_str_seconds_params: list[tuple[timedelta, str]] = [
     (timedelta(hours=2, minutes=8, seconds=6, milliseconds=13), "2h 8m"),
